Fix circular dependency in prettyhtml-hast-to-html
The "all" module depended on the "one" module and vice versa.
Fix this by defining the functionality needed by "one" (a set of
handlers) in a seperate object which we can pass around.

This circular dependency could cause the library to fail in some
cases, depending on how it was bundled.